1. fix issues; 2. reorg modules
